requirements.txt
async-timeout==5.0.1
cffi==1.17.1
click==8.2.1
cryptography==45.0.3
h11==0.16.0
hiredis==2.4.0
iniconfig==2.1.0
more-itertools==8.14.0
packaging==25.0
pluggy==1.6.0
pptree==3.1
pycparser==2.22
pydantic==1.10.22
pytest==7.3.1
python-dotenv==1.1.0
python-ulid==1.1.0
redis==4.5.1
redis-om==0.1.2
types-cffi==1.17.0.20250523
types-pyOpenSSL==24.1.0.20240722
types-redis==4.6.0.20241004
types-setuptools==80.9.0.20250529
typing_extensions==4.13.2
uvicorn==0.34.2
